Enable selective stacking
