+++ src/noisepy/seis/noise_module.py
@@ -630,6 +630,8 @@ def stacking(cc_array, cc_time, cc_ngood, stack_para):
             allstacks1 = np.mean(cc_array, axis=0)
             allstacks2 = pws(cc_array, sampling_rate)
             allstacks3, w, nstep = robust_stack(cc_array, 0.001)
+        elif smethod == StackMethod.SELECTIVE:
+            allstacks1, cc = selective_stack(cc_array, 0.001)
         nstacks = np.sum(cc_ngood)
 
     # good to return
@@ -1336,3 +1338,27 @@ def nroot_stack(cc_array, power):
     nstack = dout * np.abs(dout) ** (power - 1)
 
     return nstack
+
+
+def selective_stack(cc_array, epsilon):
+    """
+    this is a selective stacking algorithm developed by Jared Bryan.
+
+    PARAMETERS:
+    ----------------------
+    cc_array: numpy.ndarray contains the 2D cross correlation matrix
+    epsilon: residual threhold to quit the iteration
+    RETURNS:
+    ----------------------
+    newstack: numpy vector contains the stacked cross correlation
+
+    Written by Marine Denolle
+    """
+    cc = np.ones(cc_array.shape[0])
+    newstack = np.mean(cc_array, axis=0)
+    for i in range(cc_array.shape[0]):
+        cc[i] = np.sum(np.multiply(newstack, cc_array[i, :].T))
+    ik = np.where(cc >= epsilon)[0]
+    newstack = np.mean(cc_array[ik, :], axis=0)
+
+    return newstack, cc
